requirements.txt
slack-bolt
openai
yt-dlp
python-dotenv
python-docx
reportlab
requests
